+++ lib/components/fo_dropdown_list/fo_dropdown_option_component.dart
@@ -3,11 +3,11 @@ import '../fo_icon/fo_icon_component.dart';
 import 'fo_dropdown_option.dart';
 
 @Component(
-  selector: 'fo-dropdown-option',
-  templateUrl: 'fo_dropdown_option_component.html',
-  styleUrls: ['fo_dropdown_option_component.css'],
-  directives: [FoIconComponent, NgClass, NgIf],
-)
+    selector: 'fo-dropdown-option',
+    templateUrl: 'fo_dropdown_option_component.html',
+    styleUrls: ['fo_dropdown_option_component.css'],
+    directives: [FoIconComponent, NgClass, NgIf],
+    changeDetection: ChangeDetectionStrategy.OnPush)
 class FoDropdownOptionComponent {
   @Input()
   FoDropdownOptionRenderable model;